[Game] GM script normalization
- Changed most of the GM scripts output to use a standardized format (for V1 commands)
- Changed Faction Id from ``uint`` to ``FactionsEnum``
- Code cleaning on GM commands
- Removed obsolete chat functions
- Removed ``BroadcastPacketToFaction`` and ``BroadcastPacketToZone`` functions
- Removed ``/dummy`` and added a dummy option to use ``/spawn npc dummy`` instead
- Added debug area to ``/teleport``

+++ AAEmu.Game/Core/Packets/G2C/SCFactionRelationListPacket.cs
@@ -24,8 +24,8 @@ public class SCFactionRelationListPacket : GamePacket
         stream.Write((byte)_relations.Length); // TODO max length 200
         foreach (var relation in _relations)
         {
-            stream.Write(relation.Id);
-            stream.Write(relation.Id2);
+            stream.Write((uint)relation.Id);
+            stream.Write((uint)relation.Id2);
             stream.Write((byte)relation.State);
             stream.Write(relation.ExpTime);
             stream.Write(0L); // type(id)
